debug

+++ lib/buffers.cpp
@@ -16,7 +16,7 @@ std::shared_ptr<compress> buffers::createDefaultCompress() {
   return c;
 }
 std::shared_ptr<memoryUsage> buffers::createDefaultMemory() {
-  std::shared_ptr<memoryUsage> c(new simpleMemoryLimit(1024 * 1024 * 4 * 1024));
+  std::shared_ptr<memoryUsage> c(new simpleMemoryLimit((long long)(1024 * 1024) *(long long)( 4 * 1024));
 
   return c;
 }
